resume_parser/requirements.txt
streamlit
pdfplumber
python-docx
requests
python-dotenv
pymysql
